Update to 7.8.0 (2359)

+++ TMessagesProj/jni/voip/webrtc/modules/audio_processing/agc2/fixed_digital_level_estimator.h
@@ -31,7 +31,7 @@ class FixedDigitalLevelEstimator {
   // kSubFramesInSample. For kFrameDurationMs=10 and
   // kSubFramesInSample=20, this means that sample_rate_hz has to be
   // divisible by 2000.
-  FixedDigitalLevelEstimator(size_t sample_rate_hz,
+  FixedDigitalLevelEstimator(int sample_rate_hz,
                              ApmDataDumper* apm_data_dumper);
 
   // The input is assumed to be in FloatS16 format. Scaled input will
@@ -43,7 +43,7 @@ class FixedDigitalLevelEstimator {
 
   // Rate may be changed at any time (but not concurrently) from the
   // value passed to the constructor. The class is not thread safe.
-  void SetSampleRate(size_t sample_rate_hz);
+  void SetSampleRate(int sample_rate_hz);
 
   // Resets the level estimator internal state.
   void Reset();
@@ -55,8 +55,8 @@ class FixedDigitalLevelEstimator {
 
   ApmDataDumper* const apm_data_dumper_ = nullptr;
   float filter_state_level_;
-  size_t samples_in_frame_;
-  size_t samples_in_sub_frame_;
+  int samples_in_frame_;
+  int samples_in_sub_frame_;
 
   RTC_DISALLOW_COPY_AND_ASSIGN(FixedDigitalLevelEstimator);
 };
